+++ src/main/java/visuals/Gui.java
@@ -44,6 +44,7 @@ public class Gui extends Application implements IGui, IChartGUI {
     private final IChartController scoreController2 = new ChartController(this);
     private final IGameController controller = new GameController(this);
     private final IScoreController scoreController = new ScoreController(this);
+    private final IUserController userController = new UserController(this);
     Stage primaryStage;
 
     @FXML
@@ -726,17 +727,17 @@ public class Gui extends Application implements IGui, IChartGUI {
         String user = name.getText();
         String userPassword = password.getText();
 
-        if (controller.isLoggedIn()) {
+        if (userController.isLoggedIn()) {
             System.out.println("Already logged in");
             return;
         }
-        if (!controller.register(user, userPassword)) {
+        if (!userController.register(user, userPassword)) {
             System.out.println("Registration failed");
             return;
         }
         paneLogin.setVisible(false);
         buttonLogout.setVisible(true);
-        labelLoggedIn.setText("Logged in as " + controller.getUsername());
+        labelLoggedIn.setText("Logged in as " + userController.getUsername());
     }
 
     @Override
@@ -780,8 +781,8 @@ public class Gui extends Application implements IGui, IChartGUI {
         String user = name.getText();
         String userPassword = password.getText();
         try {
-            controller.login(user, userPassword);
-            if (!controller.isLoggedIn()) {
+            userController.login(user, userPassword);
+            if (!userController.isLoggedIn()) {
                 System.out.println("Login failed");
                 stats.setVisible(false);
                 return;
@@ -791,7 +792,7 @@ public class Gui extends Application implements IGui, IChartGUI {
             buttonLogout.setVisible(true);
             stats.setVisible(true);
 
-            labelLoggedIn.setText("Logged in as " + controller.getUsername());
+            labelLoggedIn.setText("Logged in as " + userController.getUsername());
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -1175,7 +1176,7 @@ public class Gui extends Application implements IGui, IChartGUI {
     @FXML
     public void setButtonLogout() {
             try {
-                controller.logout();
+                userController.logout();
                 labelLoggedIn.setText("Not logged in");
                 name.clear();
                 password.clear();
